+++ packages/core/src/types/index.ts
@@ -162,6 +162,8 @@ import type JumpToSliceOptions from './JumpToSliceOptions';
 
 import type GeometryLoaderFn from './GeometryLoaderFn';
 
+import type { RenderingEngineModeType } from './RenderingEngineMode';
+
 export type {
   // config
   Cornerstone3DConfig,
@@ -315,4 +317,5 @@ export type {
   HistoryMemo,
   VoxelManager,
   RLEVoxelMap,
+  RenderingEngineModeType,
 };