+++ packages/tools/src/tools/displayTools/Labelmap/labelmapDisplay.ts
@@ -32,6 +32,7 @@ import { getPolySeg } from '../../../config';
 import { computeAndAddRepresentation } from '../../../utilities/segmentation/computeAndAddRepresentation';
 import { triggerSegmentationDataModified } from '../../../stateManagement/segmentation/triggerSegmentationEvents';
 import { defaultSegmentationStateManager } from '../../../stateManagement/segmentation/SegmentationStateManager';
+import type vtkImageSlice from '@kitware/vtk.js/Rendering/Core/ImageSlice';
 
 // 255 itself is used as preview color, so basically
 // we have 254 colors to use for the segments if we are using the preview.
@@ -205,6 +206,7 @@ async function render(
   }
 
   for (const labelmapActorEntry of labelmapActorEntries) {
+    // call the function to set the color and opacity
     _setLabelmapColorAndOpacity(
       viewport.id,
       labelmapActorEntry,
@@ -317,7 +319,7 @@ function _setLabelmapColorAndOpacity(
   }
 
   ofun.setClamping(false);
-  const labelmapActor = labelmapActorEntry.actor as unknown as vtkVolume;
+  const labelmapActor = labelmapActorEntry.actor as vtkVolume | vtkImageSlice;
 
   // @ts-ignore - fix type in vtk
   const { preLoad } = labelmapActor.get?.('preLoad') || { preLoad: null };
@@ -333,7 +335,6 @@ function _setLabelmapColorAndOpacity(
   if (renderOutline) {
     // @ts-ignore - fix type in vtk
     labelmapActor.getProperty().setUseLabelOutline(renderOutline);
-
     // @ts-ignore - fix type in vtk
     labelmapActor.getProperty().setLabelOutlineOpacity(outlineOpacity);
 
@@ -363,7 +364,7 @@ function _setLabelmapColorAndOpacity(
     }
 
     labelmapActor.getProperty().setLabelOutlineThickness(outlineWidths);
-
+    // Mark the actor as modified to ensure the changes are applied
     labelmapActor.modified();
     labelmapActor.getProperty().modified();
     labelmapActor.getMapper().modified();